requirements.txt
streamlit
flake8
crewai
langchain-community
openai
